Overrides for Exposed forms to they align
Not sure yet why the search box is not expanding to 100% of the 10 columns inside the nav bar but it does when inside a views (not exposed) directly (need to compare the HTML/CSS)
But i will figure it out.
Also hides labels on exposed views form but only in the header?
and overrides .mb-3 just for that